rename: auth 폴더명 변경
